+++ Quests/Quests.cs
@@ -239,6 +239,7 @@ namespace OvermorrowMod.Quests
         public static void Unload()
         {
             QuestList.Clear();
+            QuestTypes.Clear();
         }
 
         // utils
